backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
mysqlclient==2.1.1
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
